Fix moving fetching list of geography units to worker
