+++ django_project/frontend/src/workers/merge_geometries.js
@@ -9,7 +9,7 @@
  *     (at your option) any later version.
  *
  * __author__ = 'irwan@kartoza.com'
- * __date__ = '05/01/2024'
+ * __date__ = '28/04/2025'
  * __copyright__ = ('Copyright 2023, Unicef')
  */
 export default () => {
